Fix perspective projection matrix creation
This bug was found after a stupidly long investigation into why frustum culling was producing incorrect results. This bug had the effect that the projection virtually moved the camera backwards some distance and may have distorted the image in other ways. This was found when I realized that the point of view moved even when the camera was  just rotating; just a curious behavior at first, but found to be the root cause of the frustum culling bug.